[feature/multi-team] 팀 전환을 위한 다중 팀 구현

+++ src/main/java/swm/wbj/asyncrum/domain/whiteboard/service/WhiteboardServiceImpl.java
@@ -9,15 +9,20 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import swm.wbj.asyncrum.domain.userteam.member.entity.Member;
 import swm.wbj.asyncrum.domain.userteam.member.service.MemberService;
+import swm.wbj.asyncrum.domain.userteam.team.entity.Team;
+import swm.wbj.asyncrum.domain.userteam.team.service.TeamService;
 import swm.wbj.asyncrum.domain.whiteboard.dto.*;
 import swm.wbj.asyncrum.domain.whiteboard.entity.Whiteboard;
+import swm.wbj.asyncrum.domain.whiteboard.exception.WhiteboardNotExistsException;
 import swm.wbj.asyncrum.domain.whiteboard.repository.WhiteboardRepository;
+import swm.wbj.asyncrum.global.exception.OperationNotAllowedException;
 import swm.wbj.asyncrum.global.media.AwsService;
 import swm.wbj.asyncrum.global.type.FileType;
-import swm.wbj.asyncrum.global.type.RoleType;
 import swm.wbj.asyncrum.global.type.ScopeType;
 
-import java.io.IOException;
+import static swm.wbj.asyncrum.global.media.AwsService.WHITEBOARD_BUCKET_NAME;
+import static swm.wbj.asyncrum.global.media.AwsService.WHITEBOARD_FILE_PREFIX;
+import static swm.wbj.asyncrum.global.type.ScopeType.isTeamScope;
 
 @RequiredArgsConstructor
 @Transactional
@@ -26,20 +31,15 @@ public class WhiteboardServiceImpl implements WhiteboardService {
 
     private final WhiteboardRepository whiteboardRepository;
     private final MemberService memberService;
+    private final TeamService teamService;
     private final AwsService awsService;
-
-    private static final String WHITEBOARD_BUCKET_NAME = "whiteboards";
-    private static final String WHITEBOARD_FILE_PREFIX ="whiteboard";
 
     @Override
     public WhiteboardCreateResponseDto createWhiteboard(WhiteboardCreateRequestDto requestDto) {
-        String title = requestDto.getTitle();
+        Member currentMember = memberService.getCurrentMember();
+        Team currentTeam = teamService.getTeamWithValidation(requestDto.getTeamId(), currentMember);
 
-        if(whiteboardRepository.existsByTitle(title)) {
-            throw new IllegalArgumentException("해당 제목은 이미 사용중입니다.");
-        }
-
-        Whiteboard whiteboard = requestDto.toEntity(memberService.getCurrentMember());
+        Whiteboard whiteboard = requestDto.toEntity(currentMember, currentTeam);
         Long whiteboardId = whiteboardRepository.save(whiteboard).getId();
 
         String whiteboardFileKey = createWhiteboardFileKey(memberService.getCurrentMember().getId(), whiteboardId);
@@ -55,76 +55,51 @@ public class WhiteboardServiceImpl implements WhiteboardService {
 
     @Transactional(readOnly = true)
     @Override
+    public WhiteboardReadAllResponseDto readAllWhiteboard(Long teamId, ScopeType scope, Integer pageIndex,
+                                                          Long topId, Integer sizePerPage) {
+        Member currentMember = memberService.getCurrentMember();
+        Team currentTeam = teamService.getTeamWithValidation(teamId, currentMember);
+
+        Page<Whiteboard> whiteboardPage;
+        Pageable pageable = PageRequest.of(pageIndex, sizePerPage, Sort.Direction.DESC, "id");
+
+        if(isTeamScope(scope)) {
+            whiteboardPage = (topId == 0L) ?
+                    whiteboardRepository.findAllByTeam(currentTeam, pageable) :
+                    whiteboardRepository.findAllByTeamAndTopId(currentTeam.getId(), currentMember.getId(), topId, pageable);
+        }
+        else {
+            whiteboardPage = (topId == 0L) ?
+                    whiteboardRepository.findAllByTeamAndMember(currentTeam, currentMember, pageable) :
+                    whiteboardRepository.findAllByTeamAndMemberAndTopId(currentTeam.getId(), currentMember.getId(), topId, pageable);
+        }
+
+        return new WhiteboardReadAllResponseDto(whiteboardPage.getContent(), whiteboardPage.getPageable(), whiteboardPage.isLast());
+    }
+
+    @Transactional(readOnly = true)
+    @Override
     public WhiteboardReadResponseDto readWhiteboard(Long id) {
+        Member currentMember = memberService.getCurrentMember();
         Whiteboard whiteboard = whiteboardRepository.findById(id)
-                .orElseThrow(() -> new IllegalArgumentException("해당 화이트보드 문서가 존재하지 않습니다."));
+                .orElseThrow(WhiteboardNotExistsException::new);
+
+        if(!ownsWhiteboard(currentMember, whiteboard)) {
+            throw new OperationNotAllowedException();
+        }
 
         return new WhiteboardReadResponseDto(whiteboard);
     }
 
-    @Transactional(readOnly = true)
-    @Override
-    public WhiteboardReadAllResponseDto readAllWhiteboard(Integer pageIndex, Long topId, Integer sizePerPage) {
-        Pageable pageable = PageRequest.of(pageIndex, sizePerPage, Sort.Direction.DESC, "id");
-        Page<Whiteboard> whiteboardPage = getWhiteboardPage(topId, pageable);
-
-        return new WhiteboardReadAllResponseDto(
-                whiteboardPage.getContent(), whiteboardPage.getPageable(), whiteboardPage.isLast());
-    }
-
-    private Page<Whiteboard> getWhiteboardPage(Long topId, Pageable pageable) {
-        Page<Whiteboard> whiteboardPage;
-        if(topId == 0) {
-            whiteboardPage = whiteboardRepository.findAll(pageable);
-        }
-        else {
-            whiteboardPage = whiteboardRepository.findAllByTopId(topId, pageable);
-        }
-        return whiteboardPage;
-    }
-
-    @Transactional(readOnly = true)
-    @Override
-    public WhiteboardReadAllResponseDto readAllWhiteboard(ScopeType scope, Integer pageIndex,
-                                                          Long topId, Integer sizePerPage) {
-        Pageable pageable = PageRequest.of(pageIndex, sizePerPage, Sort.Direction.DESC, "whiteboard_id");
-        Member currentMember = memberService.getCurrentMember();
-        Page<Whiteboard> whiteboardPage = getWhiteboards(scope, topId, pageable, currentMember);
-
-        return new WhiteboardReadAllResponseDto(
-                whiteboardPage.getContent(), whiteboardPage.getPageable(), whiteboardPage.isLast());
-    }
-
-    private Page<Whiteboard> getWhiteboards(ScopeType scope, Long topId, Pageable pageable, Member currentMember) {
-        Page<Whiteboard> whiteboardPage;
-        if (scope == ScopeType.TEAM && currentMember.getTeam() != null) {
-            if(topId == 0) {
-                whiteboardPage = whiteboardRepository.findAllByTeam(
-                        currentMember.getTeam().getId(), currentMember.getId(), pageable);
-            }
-            else {
-                whiteboardPage = whiteboardRepository.findAllByTeamAndTopId(
-                        currentMember.getTeam().getId(), currentMember.getId(), topId, pageable);
-            }
-        }
-        else if (scope == ScopeType.PRIVATE || currentMember.getTeam() == null) {
-            if(topId == 0) {
-                whiteboardPage = whiteboardRepository.findAllByAuthor(currentMember.getId(), pageable);
-            }
-            else {
-                whiteboardPage = whiteboardRepository.findAllByAuthorAndTopId(currentMember.getId(), topId, pageable);
-            }
-        }
-        else {
-            throw new IllegalArgumentException("허용되지 않은 범위입니다.");
-        }
-        return whiteboardPage;
-    }
-
     @Override
     public WhiteboardUpdateResponseDto updateWhiteboard(Long id, WhiteboardUpdateRequestDto requestDto) {
+        Member currentMember = memberService.getCurrentMember();
         Whiteboard whiteboard = whiteboardRepository.findById(id)
-                .orElseThrow(() -> new IllegalArgumentException("해당 화이트보드 문서가 존재하지 않습니다."));
+                .orElseThrow(WhiteboardNotExistsException::new);
+
+        if(!ownsWhiteboard(currentMember, whiteboard)) {
+            throw new OperationNotAllowedException();
+        }
 
         whiteboard.updateTitleAndDescription(requestDto.getTitle(), requestDto.getDescription());
         whiteboard.updateScope(ScopeType.of(requestDto.getScope()));
@@ -137,18 +112,22 @@ public class WhiteboardServiceImpl implements WhiteboardService {
 
     @Override
     public void deleteWhiteboard(Long id) {
+        Member currentMember = memberService.getCurrentMember();
         Whiteboard whiteboard = whiteboardRepository.findById(id)
-                .orElseThrow(() -> new IllegalArgumentException("해당 화이트보드 문서가 존재하지 않습니다."));
+                .orElseThrow(WhiteboardNotExistsException::new);
+
+        if(!ownsWhiteboard(currentMember, whiteboard)) {
+            throw new OperationNotAllowedException();
+        }
 
         awsService.deleteFile(whiteboard.getWhiteboardFileKey(), WHITEBOARD_BUCKET_NAME);
         whiteboardRepository.delete(whiteboard);
     }
 
-    /**
-     * 화이트보드 문서 파일명 생성 (Object Key)
-     * 파일명 = "whiteboard" + Member ID + Whiteboard ID
-     *     ex) whiteboard_2342_32
-     */
+    private boolean ownsWhiteboard(Member currentMember, Whiteboard whiteboard) {
+        return whiteboard.getMember().equals(currentMember);
+    }
+
     public String createWhiteboardFileKey(Long memberId, Long whiteboardId) {
         return WHITEBOARD_FILE_PREFIX + "_" + memberId + "_" + whiteboardId + "." + FileType.TLDR.getName();
     }
